added resolver for "access"
The format of the output is almost completely self-defined and will probably change as soon a standard is established in the mycore project.

For now:
One "access" element for every ObjectD as the toplevel is parent for several "pool" elements (attribute "id" holds the pool name, e.g. "READ") which contains the MCRAccessRule.getRuleElement() output of the approriate pair (poolID,objectID).

For simplicity a "pool" entry and therefor the rule entry too are obmitted if the rule is the "dummy rule".